Fix the navbar opening, clean and simplify code
Menu icon appearing in the navbar on low screen width is now functioning
properly. Navbar and search code is simplified. Code and libs for the
unused editing interface is wiped out. No offense meant to authors: your
code is saved by git and will be reused as much as possible after the
ongoing overall refactoring.